backend/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
groq
pydantic
python-multipart
pypdf
python-dotenv
SQLAlchemy
bcrypt
PyJWT
email-validator
